automatically configure threadpool
