+++ Cargo.toml
@@ -1,21 +1,20 @@
 [package]
 name = "smol-potat"
-description = "Proc macro for smol runtime."
-license = "MIT OR Apache-2.0"
-version = "0.2.1"
-authors = ["YuWeiWu <wusyong9104@gmail.com>"]
+version = "0.1.0"
+authors = ["Daniel Eades <danieleades@hotmail.com>"]
 edition = "2018"
-readme = "README.md"
-repository = "https://github.com/wusyong/smol-potat"
-homepage = "https://github.com/wusyong/smol-potat"
-documentation = "https://docs.rs/smol-potat"
 
-[lib]
-proc-macro = true
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[workspace]
+members = [
+    ".", "smol-potat-macro"
+]
 
 [dependencies]
-syn = { version = "1.0", features = ["full"] }
-quote = "1.0"
+smol-potat-macro = {path = "smol-potat-macro"}
+num_cpus = { version = "1.13.0", optional = true }
+smol = "0.1.5"
 
-[dev-dependencies]
-smol = "0.1"
+[features]
+auto = ["smol-potat-macro/auto", "num_cpus"]
