+++ smol-potat-macro/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "smol-potat-macro"
+description = "Proc macro for smol runtime."
+license = "MIT OR Apache-2.0"
+version = "0.2.1"
+authors = ["YuWeiWu <wusyong9104@gmail.com>"]
+edition = "2018"
+readme = "README.md"
+repository = "https://github.com/wusyong/smol-potat"
+homepage = "https://github.com/wusyong/smol-potat"
+documentation = "https://docs.rs/smol-potat"
+
+[lib]
+proc-macro = true
+
+[features]
+auto = []
+
+[dependencies]
+syn = { version = "1.0", features = ["full"] }
+quote = "1.0"
+
+[dev-dependencies]
+smol = "0.1"
